# swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:7227"
    }
  ],
  "paths": {
    "/api/Loans": {
      "get": {
        "tags": [
          "Loans"
        ],
        "operationId": "Loans_GetLoans",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LoanDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Loans"
        ],
        "operationId": "Loans_CreateLoan",
        "requestBody": {
          "x-name": "loan",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Loan"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Loans/{id}": {
      "get": {
        "tags": [
          "Loans"
        ],
        "operationId": "Loans_GetLoan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Loans"
        ],
        "operationId": "Loans_UpdateLoan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "loan",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Loan"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Loans"
        ],
        "operationId": "Loans_DeleteLoan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Loans/{id}/exitstrategy": {
      "get": {
        "tags": [
          "Loans"
        ],
        "operationId": "Loans_GetExitStrategy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExitStrategyBase"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Loans"
        ],
        "operationId": "Loans_SetExitStrategy",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "strategy",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExitStrategyBase"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Loans/sellorders/all": {
      "get": {
        "tags": [
          "Loans"
        ],
        "operationId": "Loans_GetAllSellOrders",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "oneOf": [
                {
                  "nullable": true,
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/SellOrderStatus"
                    }
                  ]
                }
              ]
            },
            "x-position": 1
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "sortDir",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "asc"
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SellOrderAggDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Loans/sellorders/{orderId}/open": {
      "post": {
        "tags": [
          "Loans"
        ],
        "operationId": "Loans_OpenSellOrder",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/Loans/sellorders/{orderId}/cancel": {
      "post": {
        "tags": [
          "Loans"
        ],
        "operationId": "Loans_CancelSellOrder",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Loans/{id}/sellorders": {
      "get": {
        "tags": [
          "Loans"
        ],
        "operationId": "Loans_GetSellOrdersForLoan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SellOrder"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "LoanDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "loanId",
          "loanDate",
          "loanPeriodMonths",
          "repaymentDate",
          "status",
          "loanAmountCzk",
          "interestRate",
          "repaymentAmountCzk",
          "feesBtc",
          "transactionFeesBtc",
          "collateralBtc",
          "totalSentBtc",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "loanId": {
            "type": "string",
            "minLength": 1
          },
          "loanDate": {
            "type": "string",
            "minLength": 1
          },
          "loanPeriodMonths": {
            "type": "integer",
            "format": "int32"
          },
          "repaymentDate": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "$ref": "#/components/schemas/LoanStatus"
          },
          "loanAmountCzk": {
            "type": "number",
            "format": "decimal"
          },
          "interestRate": {
            "type": "number",
            "format": "decimal"
          },
          "repaymentAmountCzk": {
            "type": "number",
            "format": "decimal"
          },
          "feesBtc": {
            "type": "number",
            "format": "decimal"
          },
          "transactionFeesBtc": {
            "type": "number",
            "format": "decimal"
          },
          "collateralBtc": {
            "type": "number",
            "format": "decimal"
          },
          "totalSentBtc": {
            "type": "number",
            "format": "decimal"
          },
          "purchasedBtc": {
            "type": "number",
            "format": "decimal"
          },
          "createdAt": {
            "type": "string",
            "minLength": 1
          },
          "updatedAt": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "LoanStatus": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Active",
          "Closed"
        ],
        "enum": [
          "Active",
          "Closed"
        ]
      },
      "Loan": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "userId",
          "loanId",
          "loanDate",
          "repaymentDate",
          "status",
          "loanAmountCzk",
          "interestRate",
          "repaymentAmountCzk",
          "feesBtc",
          "transactionFeesBtc",
          "collateralBtc",
          "totalSentBtc",
          "purchasedBtc",
          "createdAt",
          "updatedAt",
          "sellOrders"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "string",
            "minLength": 1
          },
          "user": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "loanId": {
            "type": "string",
            "minLength": 1
          },
          "loanDate": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          },
          "loanPeriodMonths": {
            "type": "integer",
            "format": "int32"
          },
          "repaymentDate": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          },
          "status": {
            "$ref": "#/components/schemas/LoanStatus"
          },
          "loanAmountCzk": {
            "type": "number",
            "format": "decimal"
          },
          "interestRate": {
            "type": "number",
            "format": "decimal"
          },
          "repaymentAmountCzk": {
            "type": "number",
            "format": "decimal"
          },
          "feesBtc": {
            "type": "number",
            "format": "decimal"
          },
          "transactionFeesBtc": {
            "type": "number",
            "format": "decimal"
          },
          "collateralBtc": {
            "type": "number",
            "format": "decimal"
          },
          "totalSentBtc": {
            "type": "number",
            "format": "decimal"
          },
          "purchasedBtc": {
            "type": "number",
            "format": "decimal"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          },
          "sellOrders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SellOrder"
            }
          },
          "strategyJson": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "User": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "email",
          "loans",
          "createdAt",
          "updatedAt",
          "drawdownFromAth",
          "absoluteLiquidationPrice",
          "ltvPercent"
        ],
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "minLength": 1
          },
          "allocatedBtc": {
            "type": "number",
            "format": "decimal"
          },
          "loans": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Loan"
            }
          },
          "coinmateApiKey": {
            "type": "string",
            "nullable": true
          },
          "coinmateApiSecret": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          },
          "lastLoginAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "drawdownFromAth": {
            "type": "number",
            "format": "decimal"
          },
          "absoluteLiquidationPrice": {
            "type": "number",
            "format": "decimal"
          },
          "ltvPercent": {
            "type": "number",
            "format": "decimal"
          }
        }
      },
      "SellOrder": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "loanId",
          "btcAmount",
          "pricePerBtc",
          "totalCzk",
          "status",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "loanId": {
            "type": "integer",
            "format": "int32"
          },
          "loan": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Loan"
              }
            ]
          },
          "coinmateOrderId": {
            "type": "string",
            "nullable": true
          },
          "btcAmount": {
            "type": "number",
            "format": "decimal"
          },
          "pricePerBtc": {
            "type": "number",
            "format": "decimal"
          },
          "totalCzk": {
            "type": "number",
            "format": "decimal"
          },
          "status": {
            "$ref": "#/components/schemas/SellOrderStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "SellOrderStatus": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Planned",
          "Submitted",
          "PartiallyFilled",
          "Completed",
          "Cancelled",
          "Failed"
        ],
        "enum": [
          "Planned",
          "Submitted",
          "PartiallyFilled",
          "Completed",
          "Cancelled",
          "Failed"
        ]
      },
      "ExitStrategyBase": {
        "type": "object",
        "x-abstract": true,
        "additionalProperties": false
      },
...
      "SellOrderAggDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "coinmateOrderId",
          "loanReference"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "loanId": {
            "type": "integer",
            "format": "int32"
          },
          "coinmateOrderId": {
            "type": "string",
            "minLength": 1
          },
          "btcAmount": {
            "type": "number",
            "format": "decimal"
          },
          "pricePerBtc": {
            "type": "number",
            "format": "decimal"
          },
          "totalCzk": {
            "type": "number",
            "format": "decimal"
          },
          "status": {
            "$ref": "#/components/schemas/SellOrderStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "loanReference": {
            "$ref": "#/components/schemas/LoanReferenceDto"
          }
        }
      },
      "LoanReferenceDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "loanId"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "loanId": {
            "type": "string",
            "minLength": 1
          },
          "loanAmountCzk": {
            "type": "number",
            "format": "decimal"
          },
          "repaymentDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}